Remove unused submessages package, clean deprecated messages
