--- components/idenstatereader/messages/messages.go
@@ -1,1 +1,0 @@
-package messages
